requirements.txt
gputools==0.2.13
h5py==3.7.0
huggingface-hub==0.10.1
imagecodecs
imageio==2.22.2
importlib-metadata==5.0.0
kiwisolver==1.4.4
llvmlite==0.39.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
monai==1.0.0
networkx==2.8.7
numba==0.56.3
numexpr
numpy
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging
pandas==1.4.4
Pillow==9.2.0
scikit-image==0.19.3
scipy==1.9.2
stardist==0.8.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2022.10.10
timm==0.6.11
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1

